+++ src/integration/mynethersdelight/java/plus/dragons/createintegratedfarming/integration/mynethersdelight/registry/MNDBlockSpoutingBehaviors.java
@@ -26,7 +26,6 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.material.Fluids;
 import net.neoforged.neoforge.fluids.FluidStack;
 import plus.dragons.createintegratedfarming.integration.ModIntegration;